v0.2.1
Reformatted the way referenced files are handled, switching between two paths that depends on whether it's run as a jar or class (only tried in VS Code). Token and User IDs are stored as separate text files instead of being hardcoded, so it helps privatize them a little more. Last but not least, added some dependencies to Maven's pom.xml to make compiling and packaging via Maven just a little nicer.