v0.3.0
Adding basic config file functionality. Still not enough exception handling, but it gets the job done assuming 1. the config file is typed correctly and 2. the parameters entered in the terminal are of the correct type. Currently the AI_MODEL_PATH and TEXT_PATH are still hardcoded to my own. Also some logs on startup are incorrect.